requirements.txt
avro-python3==1.9.1
certifi==2019.11.28
chardet==3.0.4
confluent-kafka==1.3.0
fastavro==0.22.9
idna==2.8
kafka-python==1.4.7
requests==2.22.0
urllib3==1.25.8
